app/pyproject.toml
[tool.poetry]
name = "itam-literature"
version = "0.1.0"
description = ""
authors = ["algakz"]
readme = "README.md"
packages = [{include = "itam"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.1"
aiogram = "^2.25.1"
asyncpg = "^0.27.0"
asyncio = "^3.4.3"
jinja2 = "^3.1.2"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^2.0.5.post1"
loguru = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
